chore: move TerminateNodeReason to events
