+++ ant-node/src/networking/interface/network_event.rs
@@ -17,7 +17,10 @@ use ant_protocol::{
 use libp2p::kad::{Record, RecordKey};
 use libp2p::{Multiaddr, PeerId};
 
-use crate::networking::{Addresses, driver::event::MsgResponder};
+use crate::{
+    event::TerminateNodeReason,
+    networking::{Addresses, driver::event::MsgResponder},
+};
 
 /// Events forwarded by the underlying Network; to be used by the upper layers
 pub(crate) enum NetworkEvent {
@@ -71,13 +74,6 @@ pub(crate) enum NetworkEvent {
     NetworkWideReplication {
         keys: Vec<(NetworkAddress, ValidationType)>,
     },
-}
-
-/// Terminate node for the following reason
-#[derive(Debug, Clone)]
-pub(crate) enum TerminateNodeReason {
-    HardDiskWriteError,
-    UpnpGatewayNotFound,
 }
 
 // Manually implement Debug as `#[debug(with = "unverified_record_fmt")]` not working as expected.
@@ -157,22 +153,6 @@ impl std::fmt::Debug for NetworkEvent {
             }
             NetworkEvent::NetworkWideReplication { keys } => {
                 write!(f, "NetworkEvent::NetworkWideReplication({keys:?})")
-            }
-        }
-    }
-}
-
-impl std::fmt::Display for TerminateNodeReason {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TerminateNodeReason::HardDiskWriteError => {
-                write!(f, "HardDiskWriteError")
-            }
-            TerminateNodeReason::UpnpGatewayNotFound => {
-                write!(
-                    f,
-                    "UPnP gateway not found. Enable UPnP on your router to allow incoming connections or manually port forward."
-                )
             }
         }
     }
